Add files via upload
MiniMedic is an application that can sense and digitalize human motion without any tools except one or more cell phones and a computer. There is no need to buy anything (as mobiles are everywhere) or install any software.

SinaLang is a set of projects including Sina and Sina++ Programming Languages. It can be proved that these languages are as powerful as other modern programming languages, mathematically.

Sina Distributed Computer is the second phase of the SinaLang project in which a new modern programming language called Sina# is developed in order to solve mathematical problems. Then, It is used to develop an application to solve computational and mathematical problems in a distributed and secured method.